requirements.txt
# CLI
click>=8.0.0
rich>=13.0.0

# Flickr API + OAuth
requests>=2.28.0
requests-oauthlib>=1.3.1

# Fuzzy text matching
rapidfuzz>=3.0.0

# (Optional) Image handling for perceptual hash or AI dedupe
Pillow>=9.0.0
imagehash>=4.3.1

# (Optional, for AI/CLIP matching, only if you want to experiment)
torch>=2.0.0 
# platform_system != "Darwin" or platform_machine == "arm64"
transformers>=4.25.0
tqdm>=4.65.0
